cleanup + extension name change
